Refactor: movido a animação do bottom-card pra pasta hooks e adicionado mensagem de clique confirmado no centralizar botao
