+++ hooks/useBottomCardAnimation.jsx
